shouldDelete in one position in leoCommands.py has escaped in master and now when synced with master that one occurrence needed to be restored.

+++ leo/core/leoCommands.py
@@ -2342,7 +2342,7 @@ class Commands(object):
                 if shouldDelete: p.deleteAllChildren()
                 at.read(p, force=True, atShadow=True)
             elif word == '@edit':
-                p.deleteAllChildren()
+                if shouldDelete: p.deleteAllChildren()
                 at.readOneAtEditNode(fn, p)
             else:
                 g.es_print('can not refresh from disk\n%r' % p.h)
